Rename to governor-delegator
