+++ src/strategies/governor-delegator/examples.json
@@ -2,7 +2,7 @@
   {
     "name": "Example query",
     "strategy": {
-      "name": "ens-delegator",
+      "name": "governor-delegator",
       "params": {
         "delegate": "0xb8c2c29ee19d8307cb7255e1cd9cbde883a267d5",
         "address": "0xc18360217d8f7ab5e7c516566761ea12ce7f9d72",
